permits/tax/workcontract - move details components to components folder
