+++ apps/af-features/src/pages/profile/employment/income-tax.page.tsx
@@ -1,14 +1,12 @@
 import { useEffect, useState } from 'react';
-import veroLogo from '@shared/images/logo-vero.svg';
-import { format, getYear, parseISO } from 'date-fns';
 import { Text } from 'suomifi-ui-components';
 import { IncomeTax } from '@shared/types';
 import AuthSentry from '@shared/components/auth-sentry';
 import Page from '@shared/components/layout/page';
 import CustomHeading from '@shared/components/ui/custom-heading';
-import CustomImage from '@shared/components/ui/custom-image';
 import Loading from '@shared/components/ui/loading';
 import PageSideNavLayout from '../components/profile-side-nav-layout';
+import TaxDetails from '../components/tax-details';
 
 const sleep = () => new Promise(resolve => setTimeout(resolve, 1000));
 
@@ -23,54 +21,6 @@ const DATA = {
 async function MOCK_DATA() {
   await sleep();
   return DATA;
-}
-
-function formatEuro(num) {
-  return new Intl.NumberFormat('de-DE', {
-    style: 'currency',
-    currency: 'EUR',
-  }).format(num);
-}
-
-function capitalizeFirstLetter(string) {
-  return string.charAt(0).toUpperCase() + string.slice(1);
-}
-
-function TaxInfo({ data }: { data: typeof DATA | undefined }) {
-  if (!data) {
-    return <Text>No tax information found.</Text>;
-  }
-
-  return (
-    <div className="flex flex-col gap-4">
-      <Text>
-        Tax payer type:{' '}
-        <span className="block font-semibold">
-          {capitalizeFirstLetter(data.taxPayerType)}
-        </span>
-      </Text>
-      <Text>
-        Tax rate:{' '}
-        <span className="block font-semibold">
-          {data.withholdingPercentage} %
-        </span>
-      </Text>
-      <Text>
-        Income limit for the year {getYear(new Date(data.validityDate))}:{' '}
-        <span className="block font-semibold">
-          {formatEuro(data.incomeLimit)}
-        </span>
-      </Text>
-      <Text>
-        Validity:{' '}
-        <span className="block font-semibold">
-          {format(parseISO(data.validityDate), 'dd.MM.yyyy')}
-        </span>
-      </Text>
-      <Text>Issued by Vero</Text>
-      <CustomImage src={veroLogo} alt="Vero" />
-    </div>
-  );
 }
 
 export default function IncomeTaxPage() {
@@ -110,7 +60,7 @@ export default function IncomeTaxPage() {
               vitae nunc.
             </Text>
 
-            {isLoading ? <Loading /> : <TaxInfo data={data} />}
+            {isLoading ? <Loading /> : <TaxDetails data={data} />}
           </div>
         </Page.Block>
       </PageSideNavLayout>
